Use systemctl daemon-reload in handler
This is a bit of a silly workaround around an Ansible issue. We'd
really like to include handlers substituting the service manager name
(like we do for tasks), but that only works when handlers are included
statically, which requires a configuration modification
(handler_includes_static = True in ansible.cfg). Without that
modification, using the systemd module in a handler that is always
included causes things to break in Ansible versions that don't have
systemd (such as 2.1).

So instead, in the handler and only in the handler, run "systemctl
daemon-reload" with command module instead.