Enable SAML support
For SP authentication against the IdP, just reuse the certificate that
the frontend also uses for HTTPS. Also, apply the same logic: if an
SSL key is defined, deploy it and its corresponding certificate. If
there is none, then create a self-signed cert.

Note that when using self-signed certificates, the common name (CN)
defaults to the frontend hostname. This means that if you define an
SSL key and cert for your frontend host (for example, in a group_vars
file for taiga-front), and none for your backend, then the backend
will get a self-signed cert that is named to match your frontend. This
is the "right" thing to do in terms of making the SP end of SAML
authentication work, but might trip you up in case your IdP enforces a
specific certificate chain.