Edit placeholder mapping dictionary if taiga_saml_mapping is unset
The SAML authentication backend is apparently fine with having neither
the 'id' mapping nor any of the 'attributes' children, but it does
require for there to be a "mapping" dictionary with an "attributes"
key. So, rather than put the onus on the user to populate
taiga_saml_mapping with an empty attributes key, just inject a
skeleton dictionary into settings/local.py.