Never mark the timestamp task "changed"
We might want to know when a no-op playbook run was a no-op playbook
run. "command" and "shell" tasks always return "changed" unless
guarded with a "creates" argument, which of course doesn't make sense
for a task that merely exists to produce a timestamp.

So, just override with "changed_when: false".