Enforce minimum memory requirement
The Taiga production setup guide explicitly states a 1GB minimum
memory requirement, "needed for compilation of lxml". Enforce this
requirement on taiga-back through the taiga_back_required_memory
variable, which is given in MB and defaults to 1024.

On low-memory platforms such as the Raspberry Pi 3 (1 GB nominal total
memory, 912 MB usable), it is still possible to compile and install
lxml, under the following conditions (at least for the duration of the
compilation):

1. Swap space must be available;
2. PostgreSQL must be stopped (this must be done manually, the role
   does not have a task for this);
3. RabbitMQ must not run (can be accomplished by setting
   taiga_enable_events and taiga_enable_async_tasks to false; both
   those features are probably not of much use on a Raspberry Pi
   anyway).

With all these hacks in place, the memory requirement can then be
relaxed by setting taiga_back_required_memory to a value like 900.

Reference:
https://taigaio.github.io/taiga-doc/dist/setup-production.html#_overview