Only make taiga-front install node.js if configured to run with gulp
If the taiga-front server uses a "-dist" frontend, then there is no
need for it to use gulp, and hence no need to install node.js on that
server. Thus, make the taiga-front to taiga-node role dependency
conditional on the frontend installation type.

Note that the taiga-front server will of course still install node.js
if it is also a taiga-events server.