Drop unnecessary variable check
Support for the systemd module was added in Ansible 2.2; the "check
for required variables" task thus verified that taiga_service_manager
was set to "circus* if the playbook was run by Ansible 2.1 or
earlier. The README, however, has stated for a while that Ansible 2.5
is the minimal supported version, so this test is really rather
useless.