Use owner with template, rather than become with an unprivileged user
When connecting to a host as an unprivileged user, using "become"
to become *another* unprivileged user can be problematic. Since
we were using the unprivileged taiga_user in two instances only
so we could drop a file as that user, become root instead and let
it drop the file as owned by taiga_user instead.

Reference:
https://docs.ansible.com/ansible/latest/user_guide/become.html#becoming-an-unprivileged-user