Mention that upstream now recommends systemd, not circus
Upstream changed its recommendation from using Circus for service
management to systemd a while ago. Retain the taiga_service_manager
default of "circus" for backward compatibility (we don't want users
running on Circus to be switched over involuntarily), but update the
defaults file and variables.md to contain a strong recommendation to
override this.