Make PYTHONPATH for circus universal
Currently PYTHONPATH is hardcoded and assumption about
python version is made. While its recommended to use systemd,
we still can fix circus by retrieving venv lib path.
As a package location we pick gunicorn as smth we expect to
be present anyway. As a result we will get
{{ lib_path }}/gunicorn/__init__.py so we apply dirname twice to get
just lib_path that can be re-used later on